Move linters around
