--- terraform/linter/outputs.go
@@ -1,31 +1,0 @@
-package linter
-
-import (
-	"github.com/hashicorp/terraform/config"
-	"github.com/pkg/errors"
-)
-
-// LintOutputs lints outputs.
-func LintOutputs(conf *config.Config) error {
-	err := shouldNotContain(
-		conf,
-		terraformTypeData,
-		terraformTypeLocal,
-		terraformTypeModule,
-		terraformTypeProvider,
-		terraformTypeResource,
-		terraformTypeTerraform,
-		terraformTypeVariable,
-	)
-	if err != nil {
-		return err
-	}
-
-	if len(conf.Outputs) == 0 {
-		return errors.New(
-			"no outputs found, either add some or remove the file",
-		)
-	}
-
-	return nil
-}
