--- terraform/linter/resource.go
@@ -1,39 +1,0 @@
-package linter
-
-import (
-	"github.com/hashicorp/terraform/config"
-	"github.com/pkg/errors"
-)
-
-// LintResources lints resources.
-func LintResources(conf *config.Config) error {
-	err := shouldNotContain(
-		conf,
-		terraformTypeData,
-		terraformTypeOutput,
-		terraformTypeProvider,
-		terraformTypeTerraform,
-		terraformTypeVariable,
-	)
-	if err != nil {
-		return err
-	}
-
-	if len(conf.Resources) == 0 &&
-		len(conf.Modules) == 0 &&
-		len(conf.Locals) == 0 {
-		return errors.New(
-			"contains no resources, modules or locals. Either remove the file or add some resources/modules/locals",
-		)
-	}
-
-	for _, resource := range conf.Resources {
-		if resource.Mode == config.DataResourceMode {
-			return errors.New(
-				"should not contain any data resources, please move to 'data.tf'",
-			)
-		}
-	}
-
-	return nil
-}
